+++ packages/eslint-config-w3up/index.js
@@ -6,7 +6,11 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     'plugin:jsdoc/recommended',
   ],
+  parserOptions: {
+    EXPERIMENTAL_useProjectService: true,
+  },
   rules: {
+    "@typescript-eslint/no-floating-promises": "error",
     '@typescript-eslint/no-unused-vars': 'off',
     '@typescript-eslint/ban-ts-comment': 'off',