+++ packages/w3up-client/test/capability/upload.test.js
@@ -49,7 +49,7 @@ describe('StoreClient', () => {
 
       const space = await alice.createSpace('test')
       const auth = await space.createAuthorization(alice)
-      alice.addSpace(auth)
+      await alice.addSpace(auth)
       await alice.setCurrentSpace(space.did())
 
       await alice.capability.upload.add(car.roots[0], [car.cid])
@@ -103,7 +103,7 @@ describe('StoreClient', () => {
 
       const space = await alice.createSpace('test')
       const auth = await space.createAuthorization(alice)
-      alice.addSpace(auth)
+      await alice.addSpace(auth)
       await alice.setCurrentSpace(space.did())
 
       const res = await alice.capability.upload.list()
@@ -148,7 +148,7 @@ describe('StoreClient', () => {
 
       const space = await alice.createSpace('test')
       const auth = await space.createAuthorization(alice)
-      alice.addSpace(auth)
+      await alice.addSpace(auth)
       await alice.setCurrentSpace(space.did())
 
       await alice.capability.upload.remove((await randomCAR(128)).roots[0])
@@ -195,7 +195,7 @@ describe('StoreClient', () => {
 
       const space = await alice.createSpace('test')
       const auth = await space.createAuthorization(alice)
-      alice.addSpace(auth)
+      await alice.addSpace(auth)
       await alice.setCurrentSpace(space.did())
 
       const result = await alice.capability.upload.get(car.cid)